feat(web): add loading skeleton states for async data components (#603)
Audit async-data components and add layout-shaped loading skeletons with
smooth fade-in transitions and preserved error states.

- Skeleton: add WithSkeleton wrapper (error -> skeleton -> fade-in content)
- MarketGrid: replace spinner with grid of MarketCardSkeleton
- markets/[id]: replace spinner with pool-detail-shaped skeleton
- DashboardStatsSections: add isLoading prop with stat-card skeletons
- DashboardBetPanels: replace "Loading..." text with skeleton rows
- globals.css: add fade-in keyframes/utility for skeleton->content

Also fix pre-existing build/type breaks encountered along the way
(missing loggers, ClaimWinnings import, market-filtering BigInt mismatch).

+++ web/app/lib/market-filtering.ts
@@ -137,8 +137,8 @@ export function filterAndSortMarkets(
   const query = normalized.search.toLowerCase();
   const minVolume = normalizeNumberInput(normalized.minVolume);
   const maxVolume = normalizeNumberInput(normalized.maxVolume);
-  const minVolumeStroops = minVolume === null ? null : unitsToStroops(BigInt(Math.round(minVolume)));
-  const maxVolumeStroops = maxVolume === null ? null : unitsToStroops(BigInt(Math.round(maxVolume)));
+  const minVolumeStroops = minVolume === null ? null : unitsToStroops(Math.round(minVolume));
+  const maxVolumeStroops = maxVolume === null ? null : unitsToStroops(Math.round(maxVolume));
   const selectedAsset = normalized.asset.toLowerCase();
 
   const filtered = markets.filter((market) => {
